added some checks

+++ lab1WinForms/MultiLevelDocks.cs
@@ -92,6 +92,7 @@ namespace lab1WinForms
             {
                 return false;
             }
+
             using (StreamReader sr = new StreamReader(filename))
             {
                 var strs = sr.ReadLine();
@@ -143,9 +144,19 @@ namespace lab1WinForms
 
         public bool SaveLevelData(string filename, int selectedLevel)
         {
+            if (selectedLevel < 0 || selectedLevel >= countPlaces)
+            {
+                return false;
+            }
+
             if (File.Exists(filename))
             {
                 File.Delete(filename);
+            }
+
+            if (docksStages[selectedLevel] == null)
+            {
+                return false;
             }
 
             using (StreamWriter sw = new StreamWriter(filename))
@@ -176,15 +187,24 @@ namespace lab1WinForms
 
         public bool LoadLevelData(string filename, int selectedLevel)
         {
+            if (selectedLevel < 0 || selectedLevel >= countPlaces)
+            {
+                return false;
+            }
+
             if (!File.Exists(filename))
             {
                 return false;
             }
+
+
             using (StreamReader sr = new StreamReader(filename))
             {
                 var strs = sr.ReadLine();
                 if (strs.Contains("Level"))
                 {
+                    docksStages[selectedLevel].Clear();
+
                     int counter = -1;
                     ITransport warship = null;
                     while (counter < 1)
